修正: TeamAdvertiserAccountControllerIT - organizationId null を (Long) null にキャスト
AdvertiserAccountResponse の organizationId (第2引数) が null の場合、
Java コンパイラが型推論を誤り引数リスト長さ不一致エラーとなっていた。
(Long) null と明示的にキャストして解決。

+++ backend/src/test/java/com/mannschaft/app/advertising/integration/TeamAdvertiserAccountControllerIT.java
@@ -101,7 +101,7 @@ class TeamAdvertiserAccountControllerIT {
     private AdvertiserAccountResponse stubAccountResponse() {
         return new AdvertiserAccountResponse(
                 ACCOUNT_ID,
-                null,
+                (Long) null,
                 ScopeType.TEAM,
                 TEAM_ID,
                 AdvertiserAccountStatus.PENDING,
